S12 US23 deployment script updates

+++ Sprint_docs/Sprint12_docs/S12_US23_T3_ALTER_TABLE_USPS_ACS_ExtractFile_PartialMatch_et_al_backout.sql
@@ -53,8 +53,9 @@ GO
 begin tran
 go
 
-DELETE FROM [dbo].[ReceiptType]
-WHERE ReceiptType_nm = 'USPS Address Change'
+IF EXISTS(Select 1 FROM [dbo].[ReceiptType] WHERE ReceiptType_nm = 'USPS Address Change')
+	DELETE FROM [dbo].[ReceiptType]
+	WHERE ReceiptType_nm = 'USPS Address Change'
 
 commit tran
